requirements.txt
altgraph==0.17.3
argcomplete==3.1.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
msgpack==1.0.6
packaging==23.2
pluggy==1.3.0
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
pytest==7.4.2
PyYAML==6.0.1
requests==2.31.0
soupsieve==2.5
urllib3==2.0.5
wget==3.2
